requirements.txt
streamlit==1.45.0
openai==0.28.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
pandas==1.5.3
pytesseract==0.3.10
Pillow==9.5.0